Header 전역 설정, 불필요한 파일 제거
페이지마다 Header를 추가하지 않아도 됩니다
안쓰는 파일 삭제했습니다
