+++ .eslintrc.js
@@ -22,5 +22,11 @@ module.exports = {
     'react/prop-types': 0, // prop-types를 선언해주어야하는 규칙 해제
     'no-param-reassign': 0, // 파라미터 변경을 허용하지 않는 규칙 해제
     'no-alert': 0, // alert 제한 규칙 해제
+    'prettier/prettier': [
+      'error',
+      {
+        endOfLine: 'auto',
+      },
+    ],
   },
 };
